This commit starts to add functionality for defining the contexts over which plugins are called. The idea is that a plugin can be called for any one of a particular set of contexts (a "context set"). For example, a plugin with context set "SingleLung" may be called with contexts "LeftLung" or "RightLung".
Contexts exist as a hierarchy, If a result is requested for a context which is higher in the hierarchy than the contexts for which the plugin can be run, then the plugin is run for all contexts in its set and the results concatenated. For example, the LungROI context encorporates both the LeftLung and RightLung contexts. if a result is requested for the "LungROI" context, but the plugin has a context set of SingleLung, then the plugin will be run for both LeftLung and RightLung and the results concatenated to return a result for LungROI.

Conversely, if a plugin is defined for context set LungROI and a result is requested for LeftLung, then the plugin is run for LungROI, but the result is masked out to return only the LeftLung portion.


PTKContextHierarchy.m
 - Added new class to manage the relationships between contexts and context sets

PTKContextMapping.m
PTKContextSet.m
PTKContextSetMapping.m
 - Helper data structures for defining the relationships between contexts and context sets